+++ src/seismicview/main.py
@@ -11,7 +11,7 @@ async def main(
         ws_server_address: tuple[str, int],
         server_receiver_address: tuple[str, int],
         reconnect: bool = False,
-        timeout: float = 5, 
+        timeout: Optional[float] = 5,
         stop: Optional[Callable[[], bool]] = None,
         stop_listener: Optional[Callable] = None
 ):
@@ -25,6 +25,9 @@ async def main(
         use_fh = True
 
     logger = get_module_logger("Seismic View Server", CONFIG.NAME, use_file_handler=use_fh)
+    logger.info(f"Server listening in {server_receiver_address}")
+    logger.info(f"Websocket Server listening in {ws_server_address}")
+
     ws_server, messages = await start_server(ws_server_address, logger)
     server = ServerReceiver(
         address=server_receiver_address,
@@ -50,7 +53,8 @@ if __name__ == "__main__":
     try:
         asyncio.run(main(
             (CONFIG.WS_SERVER_HOST_IP, CONFIG.WS_SERVER_HOST_PORT),
-            (CONFIG.SERVER_HOST_IP, CONFIG.SERVER_HOST_PORT)
+            (CONFIG.SERVER_HOST_IP, CONFIG.SERVER_HOST_PORT),
+            timeout=None
         ))
     except KeyboardInterrupt:
         pass